requirements.txt
python-decouple

langchain
langchain-google-genai